Add deb package
The deb package is not built from this repository, but here are all of 
the "important" files to make the package work. 

Reason why I don't have the pkg public is, because of how much build 
related files it had. Bunch of patches for very small changes etc. Feel 
free to use this code to buiild you a package if you don't want to use 
the actual pkg since the source code is not public.